+++ csrc/flash_attn/src/flash_fwd_sparse_hdim192_fp16_causal_sm80.cu
@@ -2,7 +2,7 @@
 // Splitting the different head dimensions to different files to speed up compilation.
 // This file is auto-generated. See "generate_kernels.py"
 
-#include "flash_fwd_launch_template.h"
+#include "flash_fwd_sparse_launch_template.h"
 
 template<>
 void run_mha_fwd_sparse_<cutlass::half_t, 192, true>(Flash_fwd_params &params, cudaStream_t stream) {
